requirements.txt
# Logistaas Ads Intelligence System — Python Dependencies

# API clients
anthropic>=0.40.0
hubspot-api-client>=9.0.0
google-ads>=24.0.0
requests>=2.32.0

# Data processing
pandas>=2.2.0
numpy>=1.26.0
pyyaml>=6.0.1

# Web framework (API server)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
aiofiles>=23.0.0

# Scheduling
apscheduler>=3.10.0

# Email delivery
sendgrid>=6.11.0

# Database
psycopg2-binary>=2.9.0

# Environment
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.9.0
